+++ packages/typespec-azure-core/src/index.ts
@@ -7,7 +7,7 @@ export * from "./decorators.js";
 export { UnionEnum, getUnionAsEnum } from "./helpers/union-enums.js";
 export * from "./lro-helpers.js";
 export * from "./rules/prevent-rest-library.js";
-export * from "./rules/use-standard-ops.js";
+export * from "./rules/use-standard-operations.js";
 export * from "./traits.js";
 export * from "./utils.js";
 export * from "./validate.js";
