+++ packages/typespec-azure-core/test/rules/use-standard-operations.test.ts
@@ -4,7 +4,7 @@ import {
   createLinterRuleTester,
 } from "@typespec/compiler/testing";
 import { beforeEach, describe, it } from "vitest";
-import { useStandardOperations } from "../../src/rules/use-standard-ops.js";
+import { useStandardOperations } from "../../src/rules/use-standard-operations.js";
 import { createAzureCoreTestRunner } from "../test-host.js";
 
 describe("typespec-azure-core: use-standard-operations rule", () => {
